% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\title{addAlpha}
\usage{
addAlpha(col, alpha = 1)
}
\value{
Returns colors with modified alpha channel
}
\description{
Add an alpha value to a colour
}
\details{
Code taken from Markus Gesmann http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html
}
\examples{
set.seed(12345)
barplot(runif(5), col=addAlpha(ninaPalette(), 0.4))
}
\seealso{
\code{\link{ninaPalette}}
}
\author{
Jens Astrom
}
