% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLogo.R
\name{addLogo}
\alias{addLogo}
\title{addLogo}
\usage{
addLogo(x = 0.85, y = 0.1)
}
\arguments{
\item{Fill}{in}
}
\value{
Fill in
}
\description{
# Adds a NINA logo to plot
}
\details{
Colors conform to the 'NINA Grafisk handbok'
}
\examples{
plot((1:10)^2, 1:10, col=NinaPalette(), cex=4, pch=16, las=1, type="n")
addLogo()
points((1:10)^2, 1:10, col=NinaPalette(), cex=4, pch=16)
}
\author{
Jens Astrom
}

