% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_gis_layer.R
\name{clip_gis_layer}
\alias{clip_gis_layer}
\title{clip_gis_layer Get an intersection from an existing PostGIS layer and a provided sf layer}
\usage{
clip_gis_layer(
  mask = NULL,
  layer_to_clip = "grunnkart_arealregnskap",
  schema = "LandCover",
  clip_geometries = TRUE,
  recalculate_area = TRUE
)
}
\arguments{
\item{mask}{A simple feature (sf) object with an id column and a geometry. Typically a polygon or a set of polygons.}

\item{layer_to_clip}{The PostGIS layer to clip from. Default is the Grunnkart for arealregnskap.}

\item{clip_geometries}{Should the returned geometries be clipped to the masking layer? If true, the first column of the masking layer is returned as row IDs, and the clipped geometries of the layer_to_clip are returned. If false, the ids in the layer_to_clip is returned as well the first column in the masking layer, together with the whole geometries in layer_to_clip that intersects at some point with the masking layer.  Default is TRUE.}

\item{recalculate_area}{Should the column "areal_m2" be recalculated after clipping the geometries? Only in effect if clip_geometries is true. Default true.}

\item{host}{host machine for the database. Default is gisdata-db.nina.no}

\item{dbname}{host database name. Default is gisdata}
}
\value{
An sf object with your initial vector column and the intersecting geometry of
}
\description{
This function is meant to return the Grunnkart for Arealregnskap for a single or set of polygons you provide. Input polygons are provided as an sf object, with a column that identifies your polygons / buffers. Ouput is a similar sf object but with the clipped geometry and ecosystem type from the Grunnkart. Other layers can be clipped by naming the layers to clip
}
\examples{
\dontrun{
NinaR::postgreSQLConnect()

small_sf <- tibble(locality = c("Semi-nat_99", "Semi-nat_100"), geom = c("POLYGON ((1082500 7775000, 1082500 7775500, 1083000 7775500, 1083000 7775000, 1082500 7775000))", "POLYGON ((1019000 7842500, 1019000 7843000, 1019500 7843000, 1019500 7842500, 1019000 7842500))")) |>
  st_as_sf(wkt = "geom", crs = 25833)

small_sf_grunnkart <- clip_gis_layer(small_sf)
}

}
