% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectGrass.R
\name{connectGrass}
\alias{connectGrass}
\title{connectGrass}
\usage{
connectGrass()
}
\arguments{
\item{Function}{accepts no params}
}
\description{
Connects to GRASS on the Linux server at NINA.
}
\details{
Only works when run on the Linux server, and stops if checks fail to identify the running machine as such.
The function creates uses a Grass location based on the formula "u_USERNAME" where "USERNAME" is identified internally,
as the user that executes the function. This location is created if needed.
}
\examples{
\dontrun{
connectGrass()
}
}
\seealso{
\code{\link{getGrassDailyTemp}}
}
\author{
Stefan Blumentrath, Jens Astrom
}
