% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassConnect.R
\name{grassConnect}
\alias{grassConnect}
\title{grassConnect}
\usage{
grassConnect(location = "ETRS_33N", mapset = "user")
}
\arguments{
\item{Function}{accepts no params}
}
\description{
Connects to GRASS on the Linux servers at NINA.
}
\details{
Only works when run on the Linux servers, and stops if checks fail to identify the running machine as such.
The function creates uses a Grass location based on the formula "u_USERNAME" where "USERNAME" is identified internally,
as the user that executes the function. This location is created if needed.
}
\examples{
\dontrun{
grassConnect()
}
}
\seealso{
\code{\link{grassDailyTemp}}
}
\author{
Stefan Blumentrath, Jens Astrom, Bernardo Niebuhr
}
