% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grassMonthlyTemp.R
\name{grassMonthlyTemp}
\alias{grassMonthlyTemp}
\title{grassMonthlyTemp}
\usage{
grassMonthlyTemp(
  points,
  start_time,
  end_time,
  where = NULL,
  type = c("mean", "min", "max", "range")
)
}
\arguments{
\item{points}{Data frame containing latitude and longitude in etrs89 / utm zone 33n (EPSG:25833). In practice, this is identical to UTM33n
The data frame should have columns named "x" and "y" for the coordinates and a column named "site" with unique integer values.}

\item{start_time}{Start time for data selection, in format "YYYY-MM-DD".}

\item{end_time}{End time for data selection, in format "YYYY-MM-DD".}

\item{where}{string of SQL that is interpreted as a "WHERE" clause. Typically used to specify a subset of the data, see example.}

\item{type}{Type of data aggregation. "mean", "min", "max", "range", default is mean.}
}
\value{
Returns a data frame containing "x", "y", "site", "start" = starting point of record, "end" = end point of record (in this case none), and
"value" containing the monthly temperature values in requested aggregate form.
}
\description{
Fetch monthly average temperatures from given points
}
\details{
Only works when run on the Linux server, and stops if checks fail to identify the running machine as such.
The function retreives monthly temperature values from the GRASS database on NINSRV16. The data is distributed
by http://www.met.no, see ftp://ftp.met.no/projects/klimagrid/ for information.
}
\examples{
\dontrun{
grassConnect()
points <- data.frame("x" = 270877, "y" = 7039976, "site" = 1)
tmp <- grassMonthlyTemp(points = points, start_time = "2014-01-01", end_time = "2014-12-31")

##Get data just for first 6 months
tmp2 <- grassMonthlyTemp(points = points, start_time = "2014-01-01", end_time = "2014-12-31", where="strftime('\%m', start_time) <= '06'")

}
}
\seealso{
\code{\link{grassMonthlyPrecip}, \link{grassDailyTemp}}
}
\author{
Stefan Blumentrath, Jens Astrom
}
