% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jensAnalysis.R
\name{jensAnalysis}
\alias{jensAnalysis}
\title{NINA Analysis pdf format.}
\usage{
jensAnalysis(..., keep_tex = TRUE, md_extensions = c("-autolink_bare_uris",
  "+header_attributes"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Provides an Rmarkdown template for creating an analysis, with some common features for work in the NINA environment.
 A shortcut for Jens that others are free to use as well. The template is available in Rstudio through
 `New File -> R Markdown -> From Template` but can also be retrieved as shown in the example below.

 This template requires a working \code{LaTeX} installation. For Windows computers, install "MikTeX".
 For Mac, install "MacTex". Linux machines usually have \code{LaTeX} available through their repositories. "texlive"
 is the preferred version for Ubuntu, and is installed on the ninrstudio01 server and therefore
 available through the Rstudio server there.
}
\details{
The resulting template can be processed most easily in Rstudio, but can also be run
 in plain R using the \code{rmarkdown::render} function. The function is based on similar template functions in
 the package "rticles" by RStudio, and uses their utility functions.
}
\examples{

\dontrun{
# initiate a template
library(rmarkdown)
draft("MyReport.Rmd", template = "jens_analysis", package = "NinaR")

#render an article manually
render("MyReport.Rmd", jensAnalysis.R())

}

}
\seealso{
\url{http://github.com/RStudio/rticles} for other templates, and
\url{http://miktex.or/download} for windows installation and
\url{https://tug.org/mactex/} for Mac installation of LaTeX backbone.
}
