% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaColors.R
\docType{data}
\name{nina_colors}
\alias{nina_colors}
\title{ninaColors}
\format{An object of class \code{character} of length 8.}
\usage{
nina_colors
}
\arguments{
\item{...}{Character names of ninaColors}
}
\description{
Specification for NINA coorporate colors
}
\details{
Shamelessly taken from https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
\keyword{datasets}
