% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaHtmlRapport.R
\name{ninaHtmlRapport}
\alias{ninaHtmlRapport}
\title{NINA Report format for publishing on NINA.no --EXPERIMENTAL!!.}
\usage{
ninaHtmlRapport(...)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::html_document}}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Provides an Rmarkdown template for creating a NINA Report, conforming to the
 existing NINA word template. The template is available in Rstudio through `New File ->
 R Markdown -> From Template` but can also be retrieved as shown in the example below.
}
\details{
Retreive an ISBN number and reportnumber from the NINA library and fill in the relevant fields in the
 resulting template. The resulting template can be processed most easily in Rstudio, but can also be run
 in plain R using the \code{rmarkdown::render} function. The function is based on similar template functions in
 the package "rticles" by RStudio, and uses their utility functions.
}
\examples{

\dontrun{
# initiate a template
library(rmarkdown)
draft("MyReport.Rmd", template = "nina_html_rapport", package = "NinaR")

#render an article manually
render("MyReport.Rmd", ninaHtmlRrapport.R())

}

}
\seealso{
\url{http://github.com/RStudio/rticles} for other templates, and
\url{http://miktex.or/download} for windows installation and
\url{https://tug.org/mactex/} for Mac installation of LaTeX backbone.
}
