% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaKortrapport.R
\name{ninaKortrapport}
\alias{ninaKortrapport}
\title{NINA kortrapport format.}
\usage{
ninaKortrapport(..., keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris", "+header_attributes"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Provides an Rmarkdown template for creating a NINA "kortrapport", conforming to existing NINA
 word template . The function is based on similar template functions in the package "rticles" by
 RStudio, and uses their utility functions.
 Requires a working \code{LaTeX} installation. For Windows computers, install "MikTeX".
}
\details{
Retreive an ISBN number and reportnumber from the NINA library and fill in the relevant fields in the
 resulting template. The resulting template can be processed most easily in Rstudio, but can be run
 in plain R using the \code{rmarkdown::render} function.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyArticle.Rmd", template = "nina_kortrapport", package = "NinaR")
}

}
\seealso{
\url{http://github.com/RStudio/rticles} for other templates, and
\url{http://miktex.or/download} for windows installation of LaTeX backbone.
}

