% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaMap.R
\name{ninaMap}
\alias{ninaMap}
\title{NINA Map Report format.}
\usage{
ninaMap(
  ...,
  keep_tex = TRUE,
  md_extensions = c("-autolink_bare_uris", "+header_attributes")
)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF.
Note that this argument does not control whether to keep the auxiliary
files (e.g., \file{.aux}) generated by LaTeX when compiling \file{.tex} to
\file{.pdf}. To keep these files, you may set \code{options(tinytex.clean =
FALSE)}.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition of R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Provides an Rmarkdown template for creating a NINA Map Report.
 The template is available in Rstudio through `New File ->
 R Markdown -> From Template` but can also be retrieved as shown in the example below.

 This template requires a working \code{LaTeX} installation. For Windows computers, install "MikTeX".
 For Mac, install "MacTex". Linux machines usually have \code{LaTeX} available through their repositories. "texlive"
 is the preferred version for Ubuntu, and is installed on the N16 server and therefore
 available through the Rstudio server there.
}
\examples{

\dontrun{
# initiate a template
library(rmarkdown)
draft("MyReportMap.Rmd", template = "nina_map", package = "NinaR")

#render an article manually
render("MyReportMap.Rmd", ninaMap.R())

}

}
