% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaPalette.R
\name{ninaPalette}
\alias{ninaPalette}
\title{ninaPalette}
\usage{
ninaPalette()
}
\arguments{
\item{Function}{accepts no params}
}
\value{
Returns 5 figure colors
}
\description{
# Creates a NINA figure color palette
}
\details{
Colors conform to the 'NINA Grafisk handbok'
}
\examples{
set.seed(12345)
barplot(runif(5), col=ninaPalette())
}
\author{
Jens Astrom, Kari Sivertsen
}
\seealso{
\code{\link{ninaLogoPalette}}
}

