% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_nina.R
\name{scale_fill_nina}
\alias{scale_fill_nina}
\title{scale_fill_nina}
\usage{
scale_fill_nina(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in ninaPalette}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
ggplot fill scale constructor for NINA colors
}
\details{
Shamelessly taken from https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
