% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyScandinavian.R
\encoding{UTF-8}
\name{tidyScandinavian}
\alias{tidyScandinavian}
\title{tidyScandinavian}
\usage{
tidyScandinavian(string = NULL)
}
\description{
Convert unicoe hestring characters to Scandinavian letters
}
\details{
This is a convenience function that translates Scandinavian hestringadecimal characters into proper letters.
}
\examples{
# to use
myData <- data.frame(
  "locations" = c(1, 2, 3),
  "messyNames" = c(
    "Tr\xf8ndelag",
    "\xf6stersund",
    "G\xe6len"
  )
)

myData$tidyName <- tidyScandinavian(myData$messyNames)
}
\author{
Diana Bowler, Jens Astrom
}
