% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/union_gis_layer.R
\name{union_gis_layer}
\alias{union_gis_layer}
\title{union_gis_layer Union (aggregate / combine) geometries of similar type}
\usage{
union_gis_layer(layer, column_to_join_on, id_column = NULL)
}
\arguments{
\item{layer}{An sf-layer to union / combine}

\item{column_to_join_on}{Character vector of column name to join on}

\item{id_column}{Optional character vector that specifies the identifying column, that you want to keep separate.}
}
\value{
An sf-object
}
\description{
This function combines separate geometries together, that share some common type. This is done separately for different areas, if provided, using an id_column.
}
\examples{
\dontrun{
#' NinaR::postgreSQLConnect()

small_sf <- tibble(locality = c("Semi-nat_99", "Semi-nat_100"), geom = c("POLYGON ((1082500 7775000, 1082500 7775500, 1083000 7775500, 1083000 7775000, 1082500 7775000))", "POLYGON ((1019000 7842500, 1019000 7843000, 1019500 7843000, 1019500 7842500, 1019000 7842500))")) |>
  st_as_sf(wkt = "geom", crs = 25833)

small_sf_grunnkart <- clip_gis_layer(small_sf)

unioned_grunnkart <- union_gis_layer(small_sf_grunnkart)
}
}
