% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{add_logo Add a NINA-logo to ggplots}
\usage{
add_logo(
  logo_type = "black",
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  stroke_scale = 1
)
}
\arguments{
\item{logo_type}{black or white logo? defaults to "black"}

\item{xmin}{start position of logo along x-axis}

\item{xmax}{end position of logo along x-axis}

\item{ymin}{start position of logo along y-axis}

\item{ymax}{end position of logo along y-axis}
}
\value{
adds a NINA logo in vector formats to ggplots
}
\description{
This is a ggplot version of addLogo. It inserts a svg version of the NINA logo in a ggplot. You have to tinker a bit with the xmin, xmax, ymin, ymax values, and the stroke_scale a bit to get it to fit your particular plot. So far, no alpha capabilities.
}
\examples{


p <- ggplot(tibble(x = (1:10)^2, y = 1:10), aes(x = x, y = y)) +




}
