% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\title{addAlpha}
\usage{
addAlpha(col, alpha = 1)
}
\arguments{
\item{col}{vector of any of the three kinds of \R color specifications,
    i.e., either a color name (as listed by \code{\link{colors}()}), a
    hexadecimal string of the form \code{"#rrggbb"} or
    \code{"#rrggbbaa"} (see \code{\link{rgb}}), or a positive integer
    \code{i} meaning \code{\link{palette}()[i]}.}

\item{alpha}{logical value indicating whether the alpha channel (opacity)
    values should be returned.}
}
\value{
Returns colors with modified alpha channel
}
\description{
Add an alpha value to a colour
}
\details{
Code taken from Markus Gesmann http://www.magesblog.com/2013/04/how-to-change-alpha-value-of-colours-in.html
}
\examples{
set.seed(12345)
barplot(runif(5), col=addAlpha(NinaPalette(), 0.4))
}
\author{
Jens Åström
}
\seealso{
\code{\link{NinaPalette}}
}

