% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWorkload.R
\name{checkWorkload}
\alias{checkWorkload}
\title{checkWorkload}
\usage{
checkWorkload(
  user_aggregate = TRUE,
  sort_proc_by = c("CPU", "MEM"),
  procs = 50
)
}
\arguments{
\item{user_aggregate}{Summarize the processes on the user that started it? Boolean.}

\item{procs}{The number of processes to show for the "Processes" table, when not aggregating on user. Default = 50.}
}
\value{
A list consisting of 1) he total number of cores on the machine, 2) an overview of the running processes, and 3) a summary of the memory usage. The overview of the running processes depends on `user_aggregate`
}
\description{
See the current workload on your machine. ONLY WORKS ON LINUX!
}
\details{
This function utilizes the function `top` from the package `NCmisc`, and the Linux system command `free` to gather info on the system usage.
In the "Processes" table, you can see the individual usage of the running processes. Look at the %CPU column in "Processes" to see how
busy the cores are. The number of total cores on the machine is shown in "Cores". In the "Memory" table, the "Available" column is the
thing of most interest.
}
\author{
Jens Åström
}
