% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountFolders.R
\name{mountFolders}
\alias{mountFolders}
\title{mountFolders}
\usage{
mountFolders()
}
\arguments{
\item{The}{function takes no parameters but asks for the password in an interactive prompt}
}
\description{
Mounts folders from the Nina Windows environment on Ninsrv16 or ninrstudio03
}
\details{
Only works when run on these Linux server, and stops if checks fail to identify the running machine as such.
Meant to be run on the Rstudio-server at http://ninsrv16:8787 or ninrstudio03.nina.no

The function calls the login script on the servers which mounts the folders:
\itemize{
\item Prosjekter (R:\\Prosjekter)
\item winhome (M:\\)
\item Kladd (R:\\Kladd)
\item grassdata (R:\\grassdata)
\item Maler (R:\\Maler)
}

On ninrstudio03 the folders in R:\ is available through HOME\\Mounted or HOME\\Mounts
}
\examples{
\dontrun{
mountFolders()
}
}
\author{
Jens Astrom
}
