% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mountFolders.R
\name{mountFolders}
\alias{mountFolders}
\title{mountFolders}
\usage{
mountFolders()
}
\arguments{
\item{The}{function takes no parameters but asks for the password in an interactive prompt}
}
\description{
Mounts folders from the Nina Windows environment on Ninsrv16
}
\details{
Only works when run on the Linux server, and stops if checks fail to identify the running machine as such.
The function calls the login script on NINSRV16, which mounts the folders:
\itemize{
\item Prosjekter (R:\\Prosjekter)
\item winhome (M:\\)
\item Kladd (R:\\Kladd)
\item grassdata (R:\\grassdata)
\item Maler (R:\\Maler)
}
This should enable you to reach these folders through Rstudio-server at Ninsrv16.
}
\examples{
\dontrun{
mountFolders()
}
}
\author{
Jens Astrom
}

