% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaPalette.R
\name{ninaPalette}
\alias{ninaPalette}
\title{ninaPalette}
\usage{
ninaPalette(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{choose between "main" and "logo" palettes.}
}
\value{
Returns the 5 or 3 NINA figure colors
}
\description{
Creates a NINA figure color palette
}
\details{
Colors conform to the 'NINA Grafisk handbok'
}
\examples{
set.seed(12345)
barplot(runif(5), col=ninaPalette())
}
\seealso{
\code{\link{ninaLogoPalette}}
}
\author{
Jens Astrom, Kari Sivertsens
}
