% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ninaR.R
\docType{_PACKAGE}
\name{NinaR}
\alias{NinaR}
\title{NinaR: document templates and functions for NINA}
\description{
NinaR provides helper functions for using NINA color schemes
and logo in R graphics, and provides document templates
for presentations and reports using RMarkdown syntax.
}
\details{
To learn more about NinaR, start with the vignettes:
\code{browseVignettes(package = "NinaR")}
}
