% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgreSQLConnect.R
\name{postgreSQLConnect}
\alias{postgreSQLConnect}
\title{postgreSQLConnect}
\usage{
postgreSQLConnect(
  username = "postgjest",
  password = "gjestpost",
  host = "gisdata-db.nina.no",
  dbname = "gisdata",
  ...
)
}
\arguments{
\item{username}{Defaults to "postgjest"}

\item{password}{Defaunts to "gjestpost"}

\item{host}{Database server name (DNS). Defaults to gisdata-db.nina.no}

\item{dbname}{Database name. Defaults to gisdata.}

\item{...}{Additional arguments passed to the RPostgres driver (dbConnect,PqDriver-method {RPostgres}). Use for example bigint = "integer" to avoid integer64 data types.}
}
\description{
Connects to the PostgreSQL database gisdata-db.nina.no.
}
\details{
Only works when run on the Linux server, and stops if checks fail to identify the running machine as such.
The function creates uses a Grass location based on the formula "u_USERNAME" where "USERNAME" is identified internally,
as the user that executes the function. This location is created if needed.
}
\examples{
\dontrun{
postgreSQLConnect()
}
}
\author{
Jens Astrom
}
