% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_nina.R
\name{scale_color_nina}
\alias{scale_color_nina}
\title{Color scale constructor for NINA colors}
\usage{
scale_color_nina(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  name = waiver(),
  ...
)
}
\arguments{
\item{palette}{Character name of palette in ninaPalette, "main" and "logo" represents separate officially defined palettes. Others are various combinations of these colors, most usefull for continuous scales. See \link{ninaPalette}.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{name}{The name of the legend. Defaults to waiver(), in which case the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
ggplot2 color scales for NINA colors
}
\details{
Shamelessly taken from https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
}
\examples{
require(ggplot2)
ggplot2::ggplot(mpg, ggplot2::aes(cyl, hwy)) +
ggplot2::geom_point(ggplot2::aes(colour = drv)) +
  scale_color_nina(discrete = TRUE)


ggplot2::ggplot(mpg, aes(cyl, hwy)) +
ggplot2::geom_point(aes(colour = year)) +
  scale_color_nina(discrete = FALSE,
                   palette = "lightblue-yellow",
                   name = "Year of make") +
 ylab("Highway miles per gallon") +
 xlab("Number of cylinders")

}
