% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyScandinavian.R
\name{tidyScandinavian}
\alias{tidyScandinavian}
\title{Convert unicoe hestring characters to Scandinavian letters}
\usage{
tidyScandinavian(string)
}
\description{
This is a convenience function that translates Scandinavian hestringadecimal characters into proper letters.
}
\examples{
#to use
myData <- data.frame("locations" = c(1, 2, 3),
                     "messyNames" = c("Tr\\xf8ndelag",
                                  "\\xf6stersund",
                                  "G\\xe6len"))

myData$tidyName <- tidyScandinavian(myData$messyNames)

}
